% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_functions.R
\name{format_app_params_test}
\alias{format_app_params_test}
\title{Function to convert inputs of Chagas shiny application}
\usage{
format_app_params_test(scenario_vars, global_vars, advance_vars)
}
\arguments{
\item{scenario_vars}{list of scenario-specific variables (pathway_type, test1, test2 (optionally test3, test4, test5);
each test contains test_type, label, facility_type, sens, spec, cost_test)}

\item{global_vars}{list of global user-provided variables (population, ltfu, link_treatment_low, link_treatment_high)}

\item{advance_vars}{list of variables included in advanced settings that can be user-updated (prev_chagas, cost_visit, treat_effect, avg_dalys)}
}
\value{
list
}
\description{
Function takes user inputs from Chagas pathways application and converts into appropriate
format for make_params function. This includes definining LTFU, visit status, and treatment
linkage for each test type from provided info on test facility level.
}
